services/web/requirements.txt
dash
dash-bootstrap-components
gunicorn
pycorn
pandas
numpy
pyspark
psycopg2-binary
pony
sqlalchemy
python-dotenv